+++ client_option_test.go
@@ -7,21 +7,26 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/metric/nonrecording"
 
 	"github.com/vgarvardt/gue/v4/adapter"
 )
 
 func TestWithClientID(t *testing.T) {
-	clientWithDefaultID := NewClient(nil)
+	clientWithDefaultID, err := NewClient(nil)
+	require.NoError(t, err)
 	assert.NotEmpty(t, clientWithDefaultID.id)
 
 	customID := "some-meaningful-id"
-	clientWithCustomID := NewClient(nil, WithClientID(customID))
+	clientWithCustomID, err := NewClient(nil, WithClientID(customID))
+	require.NoError(t, err)
 	assert.Equal(t, customID, clientWithCustomID.id)
 }
 
 func TestWithClientLogger(t *testing.T) {
-	clientWithDefaultLogger := NewClient(nil)
+	clientWithDefaultLogger, err := NewClient(nil)
+	require.NoError(t, err)
 	assert.IsType(t, adapter.NoOpLogger{}, clientWithDefaultLogger.logger)
 
 	logMessage := "hello"
@@ -31,7 +36,8 @@ func TestWithClientLogger(t *testing.T) {
 	// worker sets id as default logger field
 	l.On("With", mock.Anything).Return(l)
 
-	clientWithCustomLogger := NewClient(nil, WithClientLogger(l))
+	clientWithCustomLogger, err := NewClient(nil, WithClientLogger(l))
+	require.NoError(t, err)
 	clientWithCustomLogger.logger.Info(logMessage)
 
 	l.AssertExpectations(t)
@@ -45,12 +51,28 @@ func TestWithClientBackoff(t *testing.T) {
 	defaultPtr := reflect.ValueOf(DefaultExponentialBackoff).Pointer()
 	customPtr := reflect.ValueOf(customBackoff).Pointer()
 
-	clientWithDefaultBackoff := NewClient(nil)
+	clientWithDefaultBackoff, err := NewClient(nil)
+	require.NoError(t, err)
+
 	clientWithDefaultBackoffPtr := reflect.ValueOf(clientWithDefaultBackoff.backoff).Pointer()
 
 	assert.Equal(t, defaultPtr, clientWithDefaultBackoffPtr)
 
-	clientWithCustomBackoff := NewClient(nil, WithClientBackoff(customBackoff))
+	clientWithCustomBackoff, err := NewClient(nil, WithClientBackoff(customBackoff))
+	require.NoError(t, err)
+
 	assert.Equal(t, customBackoff(123), clientWithCustomBackoff.backoff(123))
 	assert.NotEqual(t, defaultPtr, customPtr)
+}
+
+func TestWithClientMeter(t *testing.T) {
+	customMeter := nonrecording.NewNoopMeterProvider().Meter("custom")
+
+	_, err := NewClient(nil)
+	require.NoError(t, err)
+
+	clientWithCustomMeter, err := NewClient(nil, WithClientMeter(customMeter))
+	require.NoError(t, err)
+
+	assert.Equal(t, customMeter, clientWithCustomMeter.meter)
 }