+++ worker.go
@@ -3,11 +3,19 @@ package gue
 import (
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"runtime"
 	"sync"
 	"time"
 
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/instrument/syncint64"
+	"go.opentelemetry.io/otel/metric/nonrecording"
+	"go.opentelemetry.io/otel/metric/unit"
+	"go.opentelemetry.io/otel/trace"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/vgarvardt/gue/v4/adapter"
@@ -61,10 +69,15 @@ type Worker struct {
 	running      bool
 	pollStrategy PollStrategy
 	pollFunc     pollFunc
+	tracer       trace.Tracer
+	meter        metric.Meter
 
 	hooksJobLocked      []HookFunc
 	hooksUnknownJobType []HookFunc
 	hooksJobDone        []HookFunc
+
+	mWorked   syncint64.Counter
+	mDuration syncint64.Histogram
 }
 
 // NewWorker returns a Worker that fetches Jobs from the Client and executes
@@ -75,7 +88,7 @@ type Worker struct {
 // WithWorkerPollInterval option.
 // The default queue is the nameless queue "", which can be overridden by
 // WithWorkerQueue option.
-func NewWorker(c *Client, wm WorkMap, options ...WorkerOption) *Worker {
+func NewWorker(c *Client, wm WorkMap, options ...WorkerOption) (*Worker, error) {
 	w := Worker{
 		interval:     defaultPollInterval,
 		queue:        defaultQueueName,
@@ -83,6 +96,8 @@ func NewWorker(c *Client, wm WorkMap, options ...WorkerOption) *Worker {
 		wm:           wm,
 		logger:       adapter.NoOpLogger{},
 		pollStrategy: PriorityPollStrategy,
+		tracer:       trace.NewNoopTracerProvider().Tracer("noop"),
+		meter:        nonrecording.NewNoopMeterProvider().Meter("noop"),
 	}
 
 	for _, option := range options {
@@ -102,7 +117,7 @@ func NewWorker(c *Client, wm WorkMap, options ...WorkerOption) *Worker {
 
 	w.logger = w.logger.With(adapter.F("worker-id", w.id))
 
-	return &w
+	return &w, w.initMetrics()
 }
 
 // Run pulls jobs off the Worker's queue at its interval. This function does
@@ -156,6 +171,7 @@ func (w *Worker) WorkOne(ctx context.Context) (didWork bool) {
 	j, err := w.pollFunc(ctx, w.queue)
 
 	if err != nil {
+		w.mWorked.Add(ctx, 1, attrJobType.String(""), attrSuccess.Bool(false))
 		w.logger.Error("Worker failed to lock a job", adapter.Err(err))
 		for _, hook := range w.hooksJobLocked {
 			hook(ctx, nil, err)
@@ -166,14 +182,23 @@ func (w *Worker) WorkOne(ctx context.Context) (didWork bool) {
 		return // no job was available
 	}
 
+	processingStartedAt := time.Now()
+	ctx, span := w.tracer.Start(ctx, "Worker.WorkOne", trace.WithAttributes(
+		attribute.String("job-type", j.Type),
+	))
+	defer span.End()
+
 	ll := w.logger.With(adapter.F("job-id", j.ID), adapter.F("job-type", j.Type))
 
 	defer func() {
 		if err := j.Done(ctx); err != nil {
+			span.RecordError(fmt.Errorf("failed to mark job as done: %w", err))
 			ll.Error("Failed to mark job as done", adapter.Err(err))
 		}
+
+		w.mDuration.Record(ctx, time.Now().Sub(processingStartedAt).Milliseconds(), attrJobType.String(j.Type))
 	}()
-	defer recoverPanic(ctx, ll, j)
+	defer recoverPanic(ctx, span, w.mWorked, ll, j)
 
 	for _, hook := range w.hooksJobLocked {
 		hook(ctx, j, nil)
@@ -183,10 +208,14 @@ func (w *Worker) WorkOne(ctx context.Context) (didWork bool) {
 
 	wf, ok := w.wm[j.Type]
 	if !ok {
+		w.mWorked.Add(ctx, 1, attrJobType.String(j.Type), attrSuccess.Bool(false))
+
+		span.RecordError(errors.New("job with unknown type"))
 		ll.Error("Got a job with unknown type")
 
 		errUnknownType := fmt.Errorf("worker[id=%s] unknown job type: %q", w.id, j.Type)
 		if err = j.Error(ctx, errUnknownType.Error()); err != nil {
+			span.RecordError(fmt.Errorf("failed to mark job as error: %w", err))
 			ll.Error("Got an error on setting an error to unknown job", adapter.Err(err))
 		}
 
@@ -198,7 +227,10 @@ func (w *Worker) WorkOne(ctx context.Context) (didWork bool) {
 	}
 
 	if err = wf(ctx, j); err != nil {
+		w.mWorked.Add(ctx, 1, attrJobType.String(j.Type), attrSuccess.Bool(false))
+
 		if jErr := j.Error(ctx, err.Error()); jErr != nil {
+			span.RecordError(fmt.Errorf("failed to mark job as error: %w", err))
 			ll.Error("Got an error on setting an error to an errored job", adapter.Err(jErr), adapter.F("job-error", err))
 		}
 
@@ -213,30 +245,57 @@ func (w *Worker) WorkOne(ctx context.Context) (didWork bool) {
 		hook(ctx, j, nil)
 	}
 
-	if err = j.Delete(ctx); err != nil {
+	err = j.Delete(ctx)
+	if err != nil {
+		span.RecordError(fmt.Errorf("failed to delete finished job: %w", err))
 		ll.Error("Got an error on deleting a job", adapter.Err(err))
 	}
 
+	w.mWorked.Add(ctx, 1, attrJobType.String(j.Type), attrSuccess.Bool(err == nil))
 	ll.Debug("Job finished")
 	return
 }
 
+func (w *Worker) initMetrics() (err error) {
+	if w.mWorked, err = w.meter.SyncInt64().Counter(
+		"gue_worker_jobs_worked",
+		instrument.WithDescription("Number of jobs processed"),
+		instrument.WithUnit(unit.Dimensionless),
+	); err != nil {
+		return fmt.Errorf("could not register mWorked metric: %w", err)
+	}
+
+	if w.mDuration, err = w.meter.SyncInt64().Histogram(
+		"gue_worker_jobs_duration",
+		instrument.WithDescription("Duration of the single locked job to be processed with all the hooks"),
+		instrument.WithUnit(unit.Milliseconds),
+	); err != nil {
+		return fmt.Errorf("could not register mDuration metric: %w", err)
+	}
+
+	return nil
+}
+
 // recoverPanic tries to handle panics in job execution.
 // A stacktrace is stored into Job last_error.
-func recoverPanic(ctx context.Context, logger adapter.Logger, j *Job) {
+func recoverPanic(ctx context.Context, span trace.Span, mWorked syncint64.Counter, logger adapter.Logger, j *Job) {
 	if r := recover(); r != nil {
 		// record an error on the job with panic message and stacktrace
 		stackBuf := make([]byte, 1024)
 		n := runtime.Stack(stackBuf, false)
 
-		buf := &bytes.Buffer{}
+		buf := new(bytes.Buffer)
 		fmt.Fprintf(buf, "%v\n", r)
 		fmt.Fprintln(buf, string(stackBuf[:n]))
 		fmt.Fprintln(buf, "[...]")
 		stacktrace := buf.String()
 
+		mWorked.Add(ctx, 1, attrJobType.String(j.Type), attrSuccess.Bool(false))
+		span.RecordError(errors.New("job panicked"), trace.WithAttributes(attribute.String("stacktrace", stacktrace)))
 		logger.Error("Job panicked", adapter.F("stacktrace", stacktrace))
+
 		if err := j.Error(ctx, stacktrace); err != nil {
+			span.RecordError(fmt.Errorf("failed to mark panicked job as error: %w", err))
 			logger.Error("Got an error on setting an error to a panicked job", adapter.Err(err))
 		}
 	}
@@ -273,6 +332,8 @@ type WorkerPool struct {
 	mu           sync.Mutex
 	running      bool
 	pollStrategy PollStrategy
+	tracer       trace.Tracer
+	meter        metric.Meter
 
 	hooksJobLocked      []HookFunc
 	hooksUnknownJobType []HookFunc
@@ -284,7 +345,7 @@ type WorkerPool struct {
 // Each Worker in the pool default to a poll interval of 5 seconds, which can be
 // overridden by WithPoolPollInterval option. The default queue is the
 // nameless queue "", which can be overridden by WithPoolQueue option.
-func NewWorkerPool(c *Client, wm WorkMap, poolSize int, options ...WorkerPoolOption) *WorkerPool {
+func NewWorkerPool(c *Client, wm WorkMap, poolSize int, options ...WorkerPoolOption) (*WorkerPool, error) {
 	w := WorkerPool{
 		wm:           wm,
 		interval:     defaultPollInterval,
@@ -293,6 +354,8 @@ func NewWorkerPool(c *Client, wm WorkMap, poolSize int, options ...WorkerPoolOpt
 		workers:      make([]*Worker, poolSize),
 		logger:       adapter.NoOpLogger{},
 		pollStrategy: PriorityPollStrategy,
+		tracer:       trace.NewNoopTracerProvider().Tracer("noop"),
+		meter:        nonrecording.NewNoopMeterProvider().Meter("noop"),
 	}
 
 	for _, option := range options {
@@ -305,8 +368,9 @@ func NewWorkerPool(c *Client, wm WorkMap, poolSize int, options ...WorkerPoolOpt
 
 	w.logger = w.logger.With(adapter.F("worker-pool-id", w.id))
 
+	var err error
 	for i := range w.workers {
-		w.workers[i] = NewWorker(
+		w.workers[i], err = NewWorker(
 			w.c,
 			w.wm,
 			WithWorkerPollInterval(w.interval),
@@ -314,13 +378,19 @@ func NewWorkerPool(c *Client, wm WorkMap, poolSize int, options ...WorkerPoolOpt
 			WithWorkerID(fmt.Sprintf("%s/worker-%d", w.id, i)),
 			WithWorkerLogger(w.logger),
 			WithWorkerPollStrategy(w.pollStrategy),
+			WithWorkerTracer(w.tracer),
+			WithWorkerMeter(w.meter),
 			WithWorkerHooksJobLocked(w.hooksJobLocked...),
 			WithWorkerHooksUnknownJobType(w.hooksUnknownJobType...),
 			WithWorkerHooksJobDone(w.hooksJobDone...),
 		)
+
+		if err != nil {
+			return nil, fmt.Errorf("could not init worker instance: %w", err)
+		}
 	}
 
-	return &w
+	return &w, nil
 }
 
 // Run runs all the Workers in the WorkerPool in own goroutines.